maistodos/api/pyproject.toml
[tool.poetry]
name = "@maistodos/api"
version = "0.1.0"
description = "maistodos challenge for senior backend vacancy"
authors = ["zejr <zejunior.py@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "@maistodos"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
